Add Desks app with initial hello world implementation
## SUMMARY
This commit introduces the initial implementation of the Desks application, integrating it with the internalbf app. It adds a simple `/desks` route that returns a basic HTML welcome page, establishing the foundational structure for future development. The `.replit` configuration was updated to include a workflow for running the Desks application, and a request handler was created in `internalbf.ts` to route `/desks` requests appropriately. Documentation files outline the implementation plan and project goals for the Desks application.

## TEST PLAN
1. **Integration Test**:
   - Access the `/desks` route and verify the returned HTML matches expectations.
   - Ensure logs are generated for the `/desks` requests.
   - Confirm other routes return their respective responses.

2. **Browser Test**:
   - Check the styling and content of the Desks welcome page on both desktop and iPad browsers.
