+++ apps/internalbf/docs/0.2/implementation-plan.md
@@ -1,0 +1,31 @@
+# InternalBF Implementation Plan - Version 0.2
+
+**Version:** 0.2
+
+## Version Summary
+
+This implementation will focus on expanding the service framework of InternalBF,
+adding standardized service registration patterns, improving operational
+capabilities with monitoring and health checks, and integrating additional
+internal services as needed.
+
+## Technical Goals
+
+- Develop a standardized service registration pattern
+- Implement service lifecycle management
+- Create a health check and monitoring system
+- Add additional internal service integrations
+- Improve error handling and logging
+
+## Components and Implementation
+
+_Detailed implementation plan to be developed after Version 0.1 is completed._
+
+## Testing Strategy
+
+_To be defined after components are specified._
+
+## Next Steps
+
+After completing this service expansion implementation, Version 0.3 will focus
+on security, authentication, and advanced service integration features.
