+++ apps/internalbf/docs/0.3/implementation-plan.md
@@ -1,0 +1,31 @@
+# InternalBF Implementation Plan - Version 0.3
+
+**Version:** 0.3
+
+## Version Summary
+
+This advanced implementation will focus on adding security and authentication to
+InternalBF, creating an admin dashboard for service monitoring, and implementing
+advanced third-party integrations to enhance the platform's capabilities.
+
+## Technical Goals
+
+- Add authentication layer for protected endpoints
+- Implement authorization controls
+- Create admin dashboard for service monitoring
+- Set up detailed logging and analytics
+- Support additional third-party service integrations
+
+## Components and Implementation
+
+_Detailed implementation plan to be developed after Version 0.2 is completed._
+
+## Testing Strategy
+
+_To be defined after components are specified._
+
+## Next Steps
+
+After completing this advanced implementation, future versions may focus on
+scalability, additional service integrations, and enhanced dashboard
+capabilities.
