+++ apps/internalbf/docs/implementation-plan.md
@@ -1,0 +1,100 @@
+# InternalBF Implementation Plan
+
+## Overview
+
+InternalBF serves as the core service platform for internal Bolt Foundry
+applications and utilities. This implementation plan outlines the technical
+approach across all versions, focusing on creating a maintainable and extensible
+system for hosting internal services.
+
+## Technical Architecture
+
+### Core Components
+
+1. **HTTP Server**: Deno-based HTTP server listening on configurable port
+2. **Router**: Request router to direct traffic to appropriate service handlers
+3. **Service Registry**: Pluggable system for registering new services
+4. **Environment Configuration**: Centralized management of environment
+   variables
+5. **Logging System**: Structured logging for operational monitoring
+
+### Technology Stack
+
+- **Runtime**: Deno 2 (TypeScript)
+- **HTTP**: Native Deno HTTP server
+- **Environment Management**: Using `get-configuration-var` package
+- **Logging**: Using Bolt Foundry logger package
+- **External Integrations**: Discord.js, Notion client
+
+## Implementation Strategy
+
+### Phase 0.1: Foundation
+
+1. **Basic Server Setup**:
+   - Implement HTTP server with request handler
+   - Create routing mechanism for dispatching requests
+   - Set up environment variable handling
+   - Implement structured logging
+
+2. **Initial Service Integration**:
+   - Implement ThanksBot Discord integration
+   - Set up Notion database connectivity
+   - Create message parsing and handling logic
+
+3. **Desks Integration**:
+   - Implement route handling for Desks application
+   - Create interface for routing requests to Desks handler
+
+### Phase 0.2: Service Expansion
+
+1. **Service Framework**:
+   - Develop standardized service registration pattern
+   - Implement service lifecycle management
+   - Create service configuration system
+
+2. **Operational Improvements**:
+   - Add health check endpoints
+   - Implement basic monitoring
+   - Create deployment scripts and documentation
+
+3. **Additional Services**:
+   - Integrate additional internal tools as needed
+   - Implement API endpoints for service management
+
+### Phase 0.3: Advanced Features
+
+1. **Security & Authentication**:
+   - Add authentication layer for protected endpoints
+   - Implement authorization controls
+   - Set up rate limiting and abuse prevention
+
+2. **Dashboard & Monitoring**:
+   - Create admin dashboard for service monitoring
+   - Implement detailed logging and analytics
+   - Set up alerting for service disruptions
+
+3. **Advanced Integration**:
+   - Support for additional third-party services
+   - Implement webhooks for external system integration
+   - Create API documentation system
+
+## Testing Strategy
+
+- **Unit Tests**: Test individual service components and utilities
+- **Integration Tests**: Test service interactions and API endpoints
+- **Load Tests**: Verify performance under expected load conditions
+- **Monitoring**: Set up continuous monitoring for production deployments
+
+## Deployment Strategy
+
+- **Development**: Local development using Deno run
+- **Testing**: Deployment to staging environment for integration testing
+- **Production**: Deployment to production environment with monitoring
+- **Rollback**: Procedures for rapid rollback in case of issues
+
+## Documentation
+
+- **Setup Guide**: Instructions for local development environment
+- **Service Documentation**: Documentation for each integrated service
+- **API Reference**: Reference for any exposed API endpoints
+- **Operational Procedures**: Guides for common operational tasks
