+++ apps/thanksbot/docs/implementation-plan.md
@@ -1,0 +1,85 @@
+# ThanksBot Implementation Plan
+
+## Overview
+
+ThanksBot is a Discord bot that captures thank-you messages and stores them in
+Notion. This implementation plan outlines the technical approach for developing
+and integrating the bot with other Bolt Foundry services.
+
+## Technical Architecture
+
+### Core Components
+
+1. **Discord Integration**: Using discord.js for Discord API interaction
+2. **Notion Integration**: Using Notion API for data storage
+3. **Message Parser**: For extracting user mentions and reasons
+4. **Response Handler**: For generating friendly responses
+
+### Technology Stack
+
+- **Runtime**: Deno 2 (TypeScript)
+- **Discord Client**: discord.js
+- **Database**: Notion (external)
+- **Environment Management**: Using `get-configuration-var` package
+- **Logging**: Using Bolt Foundry logger package
+
+## Implementation Strategy
+
+### Phase 0.1: Foundation
+
+1. **Discord Bot Setup**:
+   - Set up Discord bot with appropriate permissions
+   - Implement message listener for "#thanks" messages
+   - Create parsing logic for extracting mentions and reasons
+
+2. **Notion Integration**:
+   - Set up Notion database with appropriate schema
+   - Implement Notion client for creating entries
+   - Add error handling for API failures
+
+3. **Response Handling**:
+   - Implement friendly response messages
+   - Create format guidance for incorrect usage
+   - Add logging for operational monitoring
+
+### Phase 0.2: Enhanced Features
+
+1. **Reporting Capabilities**:
+   - Implement commands for generating reports
+   - Create leaderboard functionality
+   - Add time-based filtering
+
+2. **Administration**:
+   - Add admin commands for customization
+   - Implement permission management
+   - Create configuration options
+
+### Phase 0.3: Advanced Integration
+
+1. **Web Dashboard**:
+   - Create web interface for viewing thank-you history
+   - Implement user authentication
+   - Add visualization of statistics
+
+2. **Extended Integrations**:
+   - Add integration with other team tools
+   - Implement notification scheduling
+   - Create custom response templates
+
+## Testing Strategy
+
+- **Unit Tests**: Test message parsing and response generation
+- **Integration Tests**: Test Discord and Notion API integration
+- **User Testing**: Gather feedback from initial users
+
+## Deployment Strategy
+
+- **Development**: Local testing with Discord testing server
+- **Testing**: Deployment to staging environment
+- **Production**: Deployment to production Discord server
+
+## Documentation
+
+- **Setup Guide**: Instructions for configuring Discord bot and Notion
+- **User Guide**: Documentation for end users on how to use the bot
+- **Admin Guide**: Instructions for administrators
